Mov add forms from home page to project page
