+++ ambag-co/src/pages/ProjectPage.vue
@@ -21,19 +21,17 @@
           />
         </template>
 
-        <!-- <ProjectComponent
-          project-title="Aspiring Filmmakers Funding"
-          target-recipient="JELOTODA"
-          :goal-funding="2000"
-          background-image="/sample.jpg"
-        />
+        <q-btn
+        flat
+        round
+        icon="add"
+        text-color="teal"
+        style="background-color: #defade; position: fixed; top: 85%; left: 75%"
+        size="20px"
+        @click="openDialog"
+      />
+      <AddProject v-model="isDialogOpen" />
 
-        <ProjectComponent
-          project-title="Agriculture Funding"
-          target-recipient="FAP"
-          :goal-funding="5000"
-          background-image="/sample2.jpg"
-        /> -->
       </div>
     </q-page-container>
   </q-layout>
@@ -45,6 +43,7 @@ import HeaderLayout from '../layouts/HeaderLayout.vue';
 import ProjectComponent from '../components/ProjectComponent.vue';
 import { getProjects } from 'src/utils/firebase';
 import { Project } from '../types/Users';
+import AddProject from './AddProject.vue';
 
 const projects = ref<Project[]>([]);
 
@@ -53,4 +52,14 @@ onBeforeMount(() => {
     projects.value = data;
   });
 });
+
+
+const isDialogOpen = ref(false);
+
+console.log(isDialogOpen.value)
+const openDialog = () => {
+  isDialogOpen.value = true;
+};
+
+
 </script>
